requirements.txt
streamlit
tensorflow==2.19
keras==3.12.1
numpy
pillow